# stat/distuv: improve Gamma.Rand performances for 0.2 <= alpha < 1